refactor&style: surcharge 저장 및 Discount enum 네이밍 변경 - Discount -> AgeDiscountType - Line 엔티티 save 시 surcharge 저장 - Fare 관련 클래스 Path 패키지로 이동
