+++ src/main/java/nextstep/subway/path/domain/AgeDiscountType.java
@@ -1,9 +1,9 @@
-package nextstep.subway.line.domain;
+package nextstep.subway.path.domain;
 
 import java.util.Arrays;
 import java.util.function.Predicate;
 
-public enum Discount {
+public enum AgeDiscountType {
     ADULT(0, 0, age -> age == null || age >= 19),
     CHILD(50, 350, age -> age >= 6 && age < 13),
     TEENAGER(20, 350, age -> age >= 13 && age < 19),
@@ -13,13 +13,13 @@ public enum Discount {
     private final int discount;
     private final Predicate<Integer> ageMatchingExp;
 
-    Discount(int discountRate, int discount, Predicate<Integer> ageMatchingExp) {
+    AgeDiscountType(int discountRate, int discount, Predicate<Integer> ageMatchingExp) {
         this.discountRate = discountRate;
         this.discount = discount;
         this.ageMatchingExp = ageMatchingExp;
     }
 
-    public static Discount of(Integer age) {
+    public static AgeDiscountType of(Integer age) {
         return Arrays.stream(values())
                 .filter(s -> s.ageMatchingExp.test(age))
                 .findFirst()
