+++ src/test/java/nextstep/subway/line/domain/DiscountTest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.domain;
 
+import nextstep.subway.path.domain.AgeDiscountType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -16,8 +17,8 @@ class DiscountTest {
             "13, TEENAGER", "18, TEENAGER",
             "19, ADULT",
             "1000, ADULT"})
-    void of(int age, Discount discount) {
-        assertThat(Discount.of(age)).isEqualTo(discount);
+    void of(int age, AgeDiscountType discount) {
+        assertThat(AgeDiscountType.of(age)).isEqualTo(discount);
     }
 
     @DisplayName("연령대별 할인된 값을 계산한다.")
@@ -28,7 +29,7 @@ class DiscountTest {
             "TEENAGER, 2000, 1320",
             "ADULT, 2000, 2000",
             "ADULT, 2000, 2000"})
-    void calculate(Discount discount, int fare, int expected) {
+    void calculate(AgeDiscountType discount, int fare, int expected) {
         assertThat(discount.calculate(fare)).isEqualTo(expected);
     }
 }
