+++ src/test/java/nextstep/subway/line/domain/FareTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.line.domain;
 
-import org.junit.jupiter.api.Test;
+import nextstep.subway.path.domain.Fare;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
@@ -9,7 +9,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class FareTest {
     @ParameterizedTest
     @CsvSource({
-            "1000, 5, 0, 0", // 0세
+            "1000, 5, 0, 0",
             "10, 20, 0, 1250", "10, 17, 0, 720", "10, 6, 0, 450",
             "18, 20, 0, 1450", "18, 17, 0, 880", "18, 6, 0, 550",
             "10, 20, 900, 2150", "10, 17, 900, 1440", "10, 6, 900, 900",
